+++ config/webpack.common.js
@@ -1,5 +1,4 @@
 const webpack = require('webpack');
-const autoprefixer = require('autoprefixer');
 const ExtractTextPlugin = require('extract-text-webpack-plugin');
 const { ENV, IS_PRODUCTION, IS_DEV, APP_VERSION, TRAVIS, dir } = require('./helpers');
 const path = require('path');
@@ -44,10 +43,6 @@ module.exports = function(options = {}) {
         {
           test: /\.css/,
           loaders: [
-            ExtractTextPlugin.extract({
-              fallback: "style-loader",
-              loader: 'css-loader'
-            }),
             'to-string-loader',
             'css-loader',
             'postcss-loader?sourceMap',
@@ -56,10 +51,6 @@ module.exports = function(options = {}) {
         {
           test: /\.scss$/,
           loaders: [
-            ExtractTextPlugin.extract({
-              fallback: 'style-loader',
-              loader: 'css-loader'
-            }),
             'to-string-loader',
             'css-loader',
             'postcss-loader?sourceMap',
@@ -87,13 +78,9 @@ module.exports = function(options = {}) {
             resourcePath: 'src'
           },
           postcss: function() {
-            return [ autoprefixer ];
+            return [require('autoprefixer') ];
           }
         }
-      }),
-      new ExtractTextPlugin({
-        filename: '[name].css',
-        allChunks: true
       })
     ]
   };