+++ src/components/item/item.component.ts
@@ -1,15 +1,8 @@
 import {
   Component,
   OnInit,
-  AfterViewInit,
   Input,
-  Output,
   ViewEncapsulation,
-  ContentChild,
-  TemplateRef,
-  ViewChild,
-  EventEmitter,
-  Renderer,
   HostBinding
 } from '@angular/core';
 
@@ -27,8 +20,8 @@ import { DraggableDirective } from '../../directives/';
  */
 @Component({
   selector: 'ngx-dnd-item',
-  templateUrl: './item.component.html',
-  styleUrls: ['./item.component.scss'],
+  templateUrl: 'item.component.html',
+  styleUrls: ['item.component.scss'],
   encapsulation: ViewEncapsulation.None
 })
 export class ItemComponent implements OnInit {