+++ src/declarations.d.ts
@@ -17,15 +17,6 @@ interface ErrorConstructor {
   stackTraceLimit: number;
 }
 
-interface Drake {
-  containers: any[];
-  remove: () => void;
-  on: (s: string, f: any) => void;
-  emit: (s: string, a: any, b: any, c: any) => void;
-  dragging: boolean;
-}
-
-declare module 'dragula';
 declare module '@swimlane/ngx-ui';
 
 declare module '*.html' {